Added CS pin for display to schematic + board
The CS pin was initially left not connected intentionally, because I was led to believe that the CS pin was only needed when using multiple devices on SPI. However, for displays like the one used, the CS pin is needed for proper display initialization. DO NOT use earlier schematics and/or board files, as these do not have a properly working display connector.